@@ -1,8 +1,9 @@
 {
   "name": "isikukood",
-  "version": "2.0.1",
+  "version": "3.0.0",
   "description": "Estonian personal ID (isikukood) JavaScript module",
-  "main": "isikukood.js",
+  "main": "index.js",
+  "type": "module",
   "directories": {
     "test": "test"
   },
@@ -27,14 +28,18 @@
   },
   "homepage": "https://github.com/dknight/Isikukood-js/",
   "devDependencies": {
+    "@types/jest": "^29.5.1",
+    "@typescript-eslint/eslint-plugin": "^5.59.6",
+    "@typescript-eslint/parser": "^5.59.6",
     "esbuild": "0.14.47",
-    "eslint": "8.18.0",
-    "jest": "28.1.2"
+    "eslint": "^8.40.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.0.4"
   },
-  "jest": {
-    "verbose": true,
-    "testEnvironmentOptions": {
-      "url": "http://localhost:4000/"
-    }
+  "dependencies": {
+    "babel-jest": "^29.5.0",
+    "jest": "^29.5.0"
   }
 }
